allow specifying custom result class
Custom storage result classes may be necessary in order to change the
name of the table used.  Currently using a custom result requires
subclassing not only the result itself, but also
VersionStorage::Standard and DeploymentHandler itself.  The latter in
particular is non-trivial.  This provides a configuration option for the
Result class to use, so that no other classes need to be replaced.