add hook to rename constraints
Constraint names are set the moment add_unique_constraint is called.
For users sub-classing to change the table name, this means that the
index was still built with the original
dbix_class_deploymenthandler_versions table name.  This resulted in
a misleadingly named index and would prevent multiple versioned schemas
from being installed to the same database because of the overlapping
index names.