rename database_version => initial_version
This makes it possible for DeploymentHandler to handle changes to its
own schema.  Currently changes such as renaming the table are impossible
because the current database version is always read at startup and the
new database version must be written after the upgrade/downgrade but
there's no way to alter the table name between the two, so one will
always fail.  Adding the option to explicitly provide the database
version makes it possible to skip the initial read and thus have DH
handle a rename of its own table.